core: extract fake tests into example binaries
